Adds WIP better return types in schema functions
Issue: When running e.g. array(/*bla*/).public() the returned instance
is not of type Array anymore but of type Schema. It is unclear how
this can be resolved without a lot of code duplication. If I recall
correctly, this PR tries to add all the functions from the Schema
class directly to the subclasses. Not ideal. But did not find a better
solution, yet.