Adds [WIP] cloneWithType return type
Goal: No need for duplication of flag-changing functions. This is to
be archived through having the correct return type on the
cloneWithType function. The return type used in this commit is
supposed to type the result as the instance class it was called
with. But for some reason it does not work:
- see obj.ts: the implementation of cloneWithType does not like the
type
- see lot's of tests: optional does not seem to work
- the type itself does not compile:

src/schema/Schema.ts:63:6 - error TS2577: Return type annotation circularly references itself.

63   ): CloneWithFlagsReturn<this, Flags>;
        ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~